==> code/05_full_model_gridded.R <==
## random forest - full model

library(sp)
library(rgdal)
library(raster)
library(randomForest)
library(kernlab)
library(rgl)
library(ks)
library(sm)
library(sf)
library(tidyverse)
library(tidylog)
library(mapview)
library(caret)
getwd()
setwd("/Users/katieirving/OneDrive - SCCWRP/Documents - Katie’s MacBook Pro/git/Arroyo_toad_RB9_V2")


# Upload Data -------------------------------------------------------------

## updated and snapped pres/abs
# inshape2 = "output_data/ToadsObs_50m_Final_snapped_coords.shp"
# orig.sdata<- shapefile(inshape2) ## p/a
# head(orig.sdata)

## upload physical data

load(file = "ignore/03_RB9_grdded_data_observations.RData") # data_hyd_sf_obs - for model build
load(file = "ignore/03_RB9_grdded_data.RData") # data_hyd_sf - for prediction

class(data_hyd_sf_obs)
NewDataObs <- data_hyd_sf_obs %>%
  as.data.frame() %>%
  dplyr::select(-geometry)

length(unique(NewDataObs$COMID)) ## 318

## get path for functions
source("original_model/Current/randomForests/PARTITIONING/DATA3/Functions.R")


# Join observations and mulitcolinearlity -------------------------------------------------------
names(NewDataObsSub)

## remove nas and rearrange for easy indexing, remove ds 50 and change neg ffm values to zero - check this, do on new data if needed
all_data_obs <- NewDataObs %>%
  # select(-DS_Mag_50) %>%
  # pivot_longer(DS_Mag_90:Wet_BFL_Mag_10, names_to = "ffm", values_to = "value") %>%
  # mutate(value = ifelse(value < 0, 0, value)) %>%
  # pivot_wider(names_from = "ffm", values_from = "value") %>%
  select(COMID, PresAbs,  MRVBF.Mx:TC_042014_RB9.3_Var, DS_Mag_50:Wet_BFL_Mag_50) %>%
  drop_na()


head(all_data_obs)
names(all_data_obs)
str(all_data_obs)
cl <- MultiColinear(all_data_obs[,c(3:75)], p=0.05)
xdata <- all_data_obs[,c(3:75)]
xdata

for(l in cl) {
  cl.test <- xdata[,-which(names(xdata)==l)]
  print(paste("REMOVE VARIABLE", l, sep=": "))
  MultiColinear(cl.test, p=0.05)
}

l
# # REMOVE MULTI-COLINEAR VARIABLES
for(l in cl) { all_data_obs <- all_data_obs[,-which(names(all_data_obs)==l)] }

### not sure how this works, check and edit

names(all_data_obs)

# Random forest model -----------------------------------------------------

set.seed(234)

# NUMBER OF BOOTSTRAP REPLICATES
b=10001

ydata <- factor(all_data_obs$PresAbs, levels = c(1,0))
xdata <- all_data_obs[,c(3:43)] 

class(ydata)
length(ydata)

# PERCENT OF PRESENCE OBSERVATIONS
( dim(all_data_obs[all_data_obs$PresAbs == 1, ])[1] / dim(all_data_obs)[1] ) * 100 ## 91%

# RUN RANDOM FORESTS MODEL SELECTION FUNCTION
#Also provides variable importance and such
#It is important to look at highest class error - Look at Global OOB errors
#TEST Object is super important; Row number corresponds to paramerers, in output
( rf.model <- rf.modelSel(x=xdata, y=ydata, imp.scale="mir", ntree=b, nodesize=5) ) 

# CREATE NEW XDATA BASED ON SELECTED MODEL AND RUN FINAL RF MODEL (Model 4)
#RF runs differently when you use symbolis languate (using ~ as in an Lin. Model... use the indexing approacy [y=rf.data[,1]...]

sel.vars <- rf.model$PARAMETERS[[1]]# set to use 1 - removes all ffm in others????

rf.data <- data.frame(y=ydata, xdata[,sel.vars])	

(rf.final <- randomForest(y=rf.data[,1], x=rf.data[,2:ncol(rf.data)], ntree=b, mtry = 7,nodesize=5,
                          importance=TRUE, norm.votes=TRUE, proximity=TRUE) )
names(rf.data)


# Tuning ------------------------------------------------------------------


##  check error rates

rf.final$err.rate

oob.error.data <- data.frame(
  Trees = rep(1:nrow(rf.final$err.rate), times = 3),
  Type = rep(c("OOB", "0", "1"), each = nrow(rf.final$err.rate)),
  Error = c(rf.final$err.rate[,"OOB"],
            rf.final$err.rate[, "0"],
            rf.final$err.rate[,"1"])
)
oob.error.data 

ggplot(data=oob.error.data, aes(x=Trees, y=Error)) +
  geom_line(aes(color=Type))

## check split

oob.values <- vector(length=10)
oob.values

for(i in 1:10) {
  temp.model <- randomForest(y~., data=rf.data, mtry = i, ntree=b)
  oob.values[i] <- temp.model$err.rate[nrow(temp.model$err.rate), 1]
  
}

oob.values

### MDS scaling from proximity

# CREATE CMD SCALED DISTANCES OF RF PROXMITIES
distanceMatrix <- dist(1 - rf.final$proximity)

rf.cmd <- cmdscale(distanceMatrix, eig=TRUE, k=4, x.ret=T)  
# pa.col <- cRamp(ydata)  
# rf.cmd <- data.frame(rf.cmd$points)

mdsVarPer <- round(rf.cmd$eig/sum(rf.cmd$eig)*100, 1)
mdsVarPer

mdsValues <- rf.cmd$points

mdsData <- data.frame(Sample = rownames(mdsValues),
                      X=mdsValues[,1],
                      Y=mdsValues[,2],
                      Status = rf.data$y)

mdsData

m1 <- ggplot(data = mdsData, aes(x=X, y=Y)) +
  geom_point(aes(color=Status)) +
  theme_bw()+
  xlab(paste("MDS1 - ", mdsVarPer[1], "%", sep=" ")) +
  ylab(paste("MDS2 - ", mdsVarPer[2], "%", sep=" ")) +
  ggtitle(paste("Pres/Abs", "PROXIMITY MATRIX", sep=" - "))
m1

file.name1 <- "/Users/katieirving/Documents/Documents - Katie’s MacBook Pro/git/Arroyo_toad_RB9/Figures/05_full_model_gridded_mds_scale.jpg"
ggsave(m1, filename=file.name1, dpi=300, height=5, width=6)




# Validation --------------------------------------------------------------

## make y data compatible
rf.data.val <- rf.data %>%
  mutate(y = ifelse(y==1, "Present", "Absent")) %>%
  mutate(y = factor(y, levels = c("Present", "Absent")))

str(rf.data.val)

## split into training and testing

setsize <- floor(nrow(rf.data)*0.8)
index <- sample(1:nrow(rf.data), size = setsize)
training <- rf.data.val[index,]
testing <- rf.data.val[-index,]
training


trcontrol = trainControl(method='cv', number=10, savePredictions = T,
                         classProbs = TRUE,summaryFunction = twoClassSummary,returnResamp="all")

model = train(y ~ . , data=training, method = "rf", trControl = trcontrol,metric="ROC") 

model$resample

model

  

# mtry  ROC        Sens       Spec     
# 2    0.9521089  0.9857671  0.5961039
# 21    0.9545505  0.9857671  0.6058442
# 41    0.9448486  0.9848476  0.6008658
# 
# ROC was used to select the optimal model using the largest value.
# The final value used for the model was mtry = 21.


confusionMatrix(predict(model,testing),testing$y)

# Confusion Matrix and Statistics
# 
# Reference
# Prediction Present Absent
# Present     543     14
# Absent       14     27
# 
# Accuracy : 0.9532         
# 95% CI : (0.933, 0.9687)
# No Information Rate : 0.9314         
# P-Value [Acc > NIR] : 0.01752        
# 
# Kappa : 0.6334         
# 
# Sensitivity : 0.9749         
# Specificity : 0.6585         
# Pos Pred Value : 0.9749         
# Neg Pred Value : 0.6585         
# Prevalence : 0.9314         
# Detection Rate : 0.908          
# Detection Prevalence : 0.9314         
# 
# 'Positive' Class : Present 

### visualise trees

library(rpart)
getwd()
pdf("/Users/katieirving/OneDrive - SCCWRP/Documents - Katie’s MacBook Pro/git/Arroyo_toad_RB9_V2/Figures/01_Trees_gridded_full_model.pdf", width=25, height=15)

full_tree <- rpart(y~., method = "class", control = rpart.control(cp = 0, minsplit = 2), data = rf.data)
plot(full_tree)
text(full_tree, use.n = T)

dev.off()


# Coeficients and importance ----------------------------------------------

rf.final$importance[,1]
varImpPlot(rf.final)

rf.final$oob.times
rf.final$test

# PLOT BOOTSTRAP ERROR CONVERGENCE
plot(rf.final, main="Bootstrap Error Convergence")

# PLOT VARIABLE IMPORTANCE
p <- as.matrix(rf.final$importance)    
ord <- rev(order(p[,1], decreasing=TRUE)[1:dim(p)[1]])  
dotchart(p[ord,1], main="Scaled Variable Importance", pch=19)

# PLOT PROTOTYPES (MULTIVARIATE CLASS CENTERS)					  
# rf.p <- classCenter(xdata, ydata, rf.final$prox)
# v1=4; v2=1
# plot(xdata[,v1], xdata[,v2], pch=21, xlab=names(xdata)[v1], ylab=names(xdata)[v2],
#      bg=c("black", "grey94")[as.numeric(factor(ydata))], main="Data with Prototypes")
# points(rf.p[,v1], rf.p[,v2], pch=21, cex=2.5, bg=c("blue", "red"))					  


# Predict on all rb9 region-------------------------------------------------------

nhd <- st_read("/Users/katieirving/SCCWRP/SD Hydro Vulnerability Assessment - General/Data/SpatialData/NHD_reaches_RB9_castreamclassification.shp")
crs(nhd)
head(nhd)

#Index refers to the right column of probabilities - in this model the second column, which is probs of "1"

all_data <- data_hyd_sf[,c("COMID", "Longitude", "Latitude", sel.vars)] 
all_data <- na.omit(all_data)


pred <- predict(rf.final, all_data,filename="output_data/Current/Model1/SppProbs.img", type="prob",  index=2, 
                na.rm=TRUE, overwrite=TRUE, progress="window")


pred_df <- as.data.frame(predict(rf.final, all_data,filename="output_data/Current/Model1/SppProbs.img", type="prob",  index=2, 
                                 na.rm=TRUE, overwrite=TRUE, progress="window"))
## add comids
pred_df$COMID <- all_data$COMID
pred_df$y <- all_data$Longitude
pred_df$x <- all_data$Latitude
pred_df

# partial dependence plots ------------------------------------------------

sp=0.6
pdf("/Users/katieirving/OneDrive - SCCWRP/Documents - Katie’s MacBook Pro/git/Arroyo_toad_RB9_V2/Figures/PartialPlots_6_full_model_gridded.pdf", width=8, height=8)
p <- as.matrix(rf.final$importance)    
ord <- rev(order(p[,1], decreasing=TRUE)[1:dim(p)[1]])  
dotchart(p[ord,1], main="Scaled Variable Importance", pch=19)
plot(rf.final, main="Bootstrap Error Convergence")

names(rf.data)
for (i in 1:length(names(rf.data[,2:ncol(rf.data)])) ) {
  p <- partialPlot(rf.final, rf.data[,2:ncol(rf.data)], 
                   names(rf.data[,2:ncol(rf.data)])[i], which.class="1", plot=FALSE)   
  p$y <- (p$y - min(p$y)) / (max(p$y) - min(p$y)) 
  plot( y=lowess(y=p$y, x=p$x, f=sp)$y, x=p$x, type="l", ylab="p",  
        xlab=names(rf.data[,2:ncol(rf.data)])[i],
        main=paste("PARTIAL PLOT", names(rf.data[,2:ncol(rf.data)])[i], sep=" - ")) 
}
dev.off() 

## partial plots only hydro
# 
# pdf("/Users/katieirving/Documents/Documents - Katie’s MacBook Pro/git/Arroyo_toad_RB9/Figures/PartialPlots_hydro_fa_mag.pdf", width=8, height=8)
# 
# hyd_vars <- rf.data[,44:47]
# hyd_vars
# i= 45
# names(rf.data)
# 
# for (i in 44:length(names(hyd_vars)) ) {
#   p <- partialPlot(rf.final, rf.data[,2:ncol(rf.data)], 
#                    names(rf.data[,2:ncol(rf.data)])[i], which.class="1", plot=FALSE)   
#   p$y <- (p$y - min(p$y)) / (max(p$y) - min(p$y)) 
#   plot( y=lowess(y=p$y, x=p$x, f=sp)$y, x=p$x, type="l", ylab="p",  
#         xlab=names(rf.data[,2:ncol(rf.data)])[i],
#         main=paste("PARTIAL PLOT", names(rf.data[,2:ncol(rf.data)])[i], sep=" - ")) 
# }
# dev.off() 



# Plotting probability of occurrence --------------------------------------

nhd <- st_read("/Users/katieirving/SCCWRP/SD Hydro Vulnerability Assessment - General/Data/SpatialData/NHD_reaches_RB9_castreamclassification.shp")
crs(nhd)
head(nhd)
head(pred_df)

length(unique(pred_df$COMID))

head(nhd)

names(all_data)

obs <- all_data_obs %>%
  dplyr::select(COMID, PresAbs)

obs

nhd_lines_rb9 <- nhd %>%
  dplyr::select(COMID) %>%
  filter(COMID %in% pred_df$COMID) %>%
  mutate(COMID = as.integer(COMID))

head(nhd_lines_prob)
dim(nhd_lines_rb9)
str(nhd_lines_rb9)

nhd_lines_prob <- full_join(nhd_lines_rb9, pred_df, by = "COMID")
nhd_lines_prob <- st_zm(nhd_lines_prob)

plot(nhd_lines_prob[5])

### round prob values - change later
nhd_lines_prob <- nhd_lines_prob %>%
  rename(probOcc = 2) %>%
  mutate(probRound = round(probOcc, digits=1))

## join in obs

nhd_lines_obs <- right_join(nhd_lines_rb9, obs, by = "COMID")
nhd_lines_obs <- st_zm(nhd_lines_obs) %>% mutate(nhd_lines_obs, PresAbs = as.factor(PresAbs)) %>%
   st_centroid()

library(viridis)

library(ggmap)

nhd_lines_probx <- nhd_lines_prob %>%
  dplyr::select(COMID, probOcc, geometry) %>% st_zm()

st_write(nhd_lines_probx, "output_data/full_model/01_full_model_prob_occurrence_gridded.shp", layer_options = "OVERWRITE=true")

# For google map, you have to give the center of the window you are looking at.
# Possibility for the map type argument: terrain / satellite / roadmap / hybrid

my_breaks <- c(0, 0.25, 0.5, 0.75, 1)

map1 <- ggplot() +
  geom_sf(data = nhd_lines_prob, aes(color = probOcc)) + 
  scale_colour_gradientn(colours=inferno(6), name="Probability of Occurrence",
                         labels = my_breaks) +
  geom_sf(data = subset(nhd_lines_obs, PresAbs == 1), shape = 1, size = 2) 


map1


file.name1 <- "/Users/katieirving/Documents/Documents - Katie’s MacBook Pro/git/Arroyo_toad_RB9/Figures/01_full_model_prob_occs_map.jpg"
ggsave(map1, filename=file.name1, dpi=300, height=5, width=6)

library(mapview)
library(sf)
library(RColorBrewer)
# webshot::install_phantomjs()

## map
# set background basemaps:
basemapsList <- c("Esri.WorldTopoMap", "Esri.WorldImagery",
                  "Esri.NatGeoWorldMap",
                  "OpenTopoMap", "OpenStreetMap", 
                  "CartoDB.Positron", "Stamen.TopOSMFeatures")

mapviewOptions(basemaps=basemapsList, vector.palette = colorRampPalette(c(  "red", "green")) , fgb = FALSE)


m1 <- mapview(nhd_lines_prob, zcol = "probOcc",  legend = TRUE, layer.name = "Probability of Occurrence") +
  mapview(subset(nhd_lines_obs, NewObs == 1), col.regions = "black",cex = 2, layer.name = "Observations")

m1@map %>% leaflet::addMeasure(primaryLengthUnit = "meters")

mapshot(m1, url = paste0(getwd(), "/ignore/01_full_model_prob_occs_mapview.html"),
        file = paste0(getwd(), "/ignore/01_full_model_prob_occs_mapview.png"))
getwd()




##################################################
# PROXIMITY MULTIDIMENSIONAL SCALING (MDS) PLOTS #
##################################################

######### FUNCTION TO CREATE COLOR VECTOR #########
#AWESOME!!! SAVE THIS FOREVER!!! - Orders color ramp with data#
cRamp <- function(x,d=c("blue", "red")){
  crange <- function(x)(as.numeric(x)-min(as.numeric(x)))/diff(range(as.numeric(x)))
  cols <- colorRamp(d)(crange(x))
  apply(cols, 1, function(xt)rgb(xt[1], xt[2], xt[3], maxColorValue=255))
}  


###################################################

# CREATE CMD SCALED DISTANCES OF RF PROXMITIES
rf.cmd <- cmdscale(1 - rf.final$proximity, eig=TRUE, k=4)  
pa.col <- cRamp(ydata)  
rf.cmd <- data.frame(rf.cmd$points)

# 2D PLOT OF FIRST TWO MDS DIMENSIONS	
plot(rf.cmd[,1:2], ylab="DIM 1", xlab="DIM 2", pch=16, col=pa.col,
     main= paste("Pres/Abs", "PROXIMITY MATRIX MDS d=2", sep=" - "))
legend("bottomright", pch=c(16,16), col=c("blue", "red"), 
       legend=c("Absent","Present")) 

# 3D PLOT OF FIRST THREE MDS DIMENSIONS - Interactive using Windows RGL Driver (Library RGL)	   
plot3d(rf.cmd[,1],rf.cmd[,2],rf.cmd[,3], col=pa.col,
       pch=18, size=1.25, type="s", xlab="MDS dim 1", ylab="MDS dim 2", 
       zlab="MDS dim 3")	

# SAVE R IMAGE 
save.image( paste(path, "Cur_10x_Final_Feb2014/Model1/RFClassModel.RData", sep="/") )
